Change getAvailableTripPatterns to hook
