+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/Dashboard_TripSearchScreen/components/ResultItem.tsx
@@ -52,11 +52,10 @@ import {
   significantWaitTime,
   significantWalkTime,
   isLegFlexibleTransport,
-  BookingRequirement,
-  AvailableTripPattern,
 } from '@atb/travel-details-screens/utils';
 import {Destination} from '@atb/assets/svg/mono-icons/places';
 import {useFontScale} from '@atb/utils/use-font-scale';
+import {AvailableTripPattern, BookingRequirement} from '../../types';
 
 type ResultItemProps = {
   tripPattern: AvailableTripPattern;
