+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/Dashboard_TripSearchScreen/hooks.ts
@@ -11,6 +11,9 @@ import {StorageModelKeysEnum} from '@atb/storage';
 import {useDebugOverride} from '@atb/debug';
 import {useRemoteConfig} from '@atb/RemoteConfigContext';
 import {RemoteConfigKeys} from '@atb/remote-config';
+import {useNow} from '@atb/utils/use-now';
+import {isLegFlexibleTransport} from '@atb/travel-details-screens/utils';
+import {AvailableTripPattern, TripPatternWithKey} from '../types';
 
 export const useFindCityZoneInLocation = (
   location: Location | undefined,
@@ -110,5 +113,72 @@ export const useFlexibleTransportEgressModeEnabled = () => {
   return useFlexibleTransportDebugOverrideOrRemote(
     'use_flexible_on_egressMode',
     StorageModelKeysEnum.UseFlexibleTransportEgressModeDebugOverride,
+  );
+};
+
+function getLatestBookingDate(
+  latestBookingTime: string, // e.g. '15:16:00'
+  expectedStartTime: string, // e.g. '2023-07-14T17:56:32+02:00'
+): Date {
+  const expectedStartDate = new Date(expectedStartTime);
+  const latestBookingDate = new Date(
+    `${expectedStartTime.split('T')[0]}T${latestBookingTime}`,
+  );
+
+  if (latestBookingDate.getTime() > expectedStartDate.getTime()) {
+    latestBookingDate.setDate(latestBookingDate.getDate() - 1);
+  }
+
+  return latestBookingDate;
+}
+
+export const useAvailableTripPatterns = (
+  tripPatterns: TripPatternWithKey[],
+): AvailableTripPattern[] => {
+  const now = useNow(2500);
+  // todo: consider what if there are several flexibleTransportLegs?
+  const tripPatternsWithBookingRequirements = tripPatterns.map(
+    (tripPattern) => {
+      const firstFlexibleTransportLeg = tripPattern?.legs.find((leg) =>
+        isLegFlexibleTransport(leg),
+      );
+      const requiresBooking = !!firstFlexibleTransportLeg;
+
+      let requiresBookingUrgently = false;
+      let isTooEarly = false;
+      let isTooLate = false;
+
+      if (requiresBooking) {
+        const latestBookingDate = getLatestBookingDate(
+          firstFlexibleTransportLeg.bookingArrangements?.latestBookingTime,
+          firstFlexibleTransportLeg.expectedStartTime,
+        );
+
+        const timeDiffMilliSeconds = latestBookingDate.getTime() - now;
+        const oneHourMilliSeconds = 60 * 60 * 1000;
+        if (timeDiffMilliSeconds < 0) {
+          isTooLate = true; // currently assuming bookWhen == 'advanceAndDayOfTravel'
+        } else if (timeDiffMilliSeconds < oneHourMilliSeconds) {
+          requiresBookingUrgently = true;
+        } else if (timeDiffMilliSeconds > 7 * 24 * oneHourMilliSeconds) {
+          isTooEarly = true;
+        }
+      }
+
+      const bookingRequirement = {
+        requiresBooking,
+        requiresBookingUrgently,
+        isTooEarly,
+        isTooLate,
+      };
+      return {
+        ...tripPattern,
+        ...{bookingRequirement},
+      };
+    },
+  );
+
+  return tripPatternsWithBookingRequirements.filter(
+    (tpwbr) => !tpwbr.bookingRequirement.isTooLate,
   );
 };
