+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/types.ts
@@ -20,3 +20,14 @@ export type TripSearchScreenParams = {
   searchTime?: SearchTime;
   callerRoute?: {name: string};
 };
+
+export type BookingRequirement = {
+  requiresBooking: boolean;
+  requiresBookingUrgently: boolean;
+  isTooEarly?: boolean;
+  isTooLate?: boolean;
+};
+
+export type AvailableTripPattern = TripPatternWithKey & {
+  bookingRequirement: BookingRequirement;
+};
