+++ src/travel-details-screens/utils.ts
@@ -179,85 +179,3 @@ function someLegsAreByTrain(tripPattern: TripPattern): boolean {
 export function isLegFlexibleTransport(leg: Leg): boolean {
   return !!leg.line?.flexibleLineType;
 }
-
-export type BookingRequirement = {
-  requiresBooking: boolean;
-  requiresBookingUrgently: boolean;
-  isTooEarly?: boolean;
-  isTooLate?: boolean;
-};
-
-// should import from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/types';?
-// or maybe export from here instead?
-type TripPatternWithKey = TripPattern & {key: string};
-
-export type AvailableTripPattern = TripPatternWithKey & {
-  bookingRequirement: BookingRequirement;
-};
-
-function getLatestBookingDate(
-  latestBookingTime: string, // e.g. '15:16:00'
-  expectedStartTime: string, // e.g. '2023-07-14T17:56:32+02:00'
-): Date {
-  const expectedStartDate = new Date(expectedStartTime);
-  const latestBookingDate = new Date(
-    `${expectedStartTime.split('T')[0]}T${latestBookingTime}`,
-  );
-
-  if (latestBookingDate.getTime() > expectedStartDate.getTime()) {
-    latestBookingDate.setDate(latestBookingDate.getDate() - 1);
-  }
-
-  return latestBookingDate;
-}
-
-export function getAvailableTripPatterns(
-  tripPatterns: TripPatternWithKey[],
-  now: number, // change to hook and use useNow() in here instead?
-): AvailableTripPattern[] {
-  // todo: consider what if there are several flexibleTransportLegs?
-  const tripPatternsWithBookingRequirements = tripPatterns.map(
-    (tripPattern) => {
-      const firstFlexibleTransportLeg = tripPattern?.legs.find((leg) =>
-        isLegFlexibleTransport(leg),
-      );
-      const requiresBooking = !!firstFlexibleTransportLeg;
-
-      let requiresBookingUrgently = false;
-      let isTooEarly = false;
-      let isTooLate = false;
-
-      if (requiresBooking) {
-        const latestBookingDate = getLatestBookingDate(
-          firstFlexibleTransportLeg.bookingArrangements?.latestBookingTime,
-          firstFlexibleTransportLeg.expectedStartTime,
-        );
-
-        const timeDiffMilliSeconds = latestBookingDate.getTime() - now;
-        const oneHourMilliSeconds = 60 * 60 * 1000;
-        if (timeDiffMilliSeconds < 0) {
-          isTooLate = true; // currently assuming bookWhen == 'advanceAndDayOfTravel'
-        } else if (timeDiffMilliSeconds < oneHourMilliSeconds) {
-          requiresBookingUrgently = true;
-        } else if (timeDiffMilliSeconds > 7 * 24 * oneHourMilliSeconds) {
-          isTooEarly = true;
-        }
-      }
-
-      const bookingRequirement = {
-        requiresBooking,
-        requiresBookingUrgently,
-        isTooEarly,
-        isTooLate,
-      };
-      return {
-        ...tripPattern,
-        ...{bookingRequirement},
-      };
-    },
-  );
-
-  return tripPatternsWithBookingRequirements.filter(
-    (tpwbr) => !tpwbr.bookingRequirement.isTooLate,
-  );
-}
